+++ test/matchers/associations_test.rb
@@ -10,6 +10,7 @@ describe "Associations" do
     it { must have_and_belong_to_many(:friends).of_type(Person) }
 
     it { must embed_one(:profile) }
+    it { must embed_many(:sites) }
   end
 
   describe Pet do