+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### 0.0.3
 
++ Added `embed_many` association matcher.
 + Fixed issue check association_type if metadata is not nil.
 + Added `embed_one(association_name)` association matcher.
 + Added `have_and_belong_to_many(association_name)` association matcher.